Just testing different values, also just first checking everything is pushed as expected this time around.
Seems navigation aps like to see a fix with at least 150 accuracy before starting navigation so am just checking out a few values. It so far seems much faster having a higher threshold and dropping down to an accuracy below 150 then it is to start looking with it set so low it ignores everything off the bat. So anyway just looking for the sweet spot ware it doesn't get caught up hogging all the satellite connections with those with not so good accuracy but at the same time not so low it doesn't take anything.