Added in “libstagefrighthw.so” from Sharp 2.3.5
For the most part because many of the libs already used list “libstagefrighthw.so” .  Also because CM7 does not use the drm overlay used by many files and adding the code drm overlay to the existing code has been difficult. Until then quality is excellent when included. Again is also part of a larger overhaul that will also use “libOmxVenc” & “libOmxVdec” so removed the lines, not that I think it makes any difference.