Edit “init.triumph.rc”
Changed permissions needed for using the Sharp 2.3.5 “hcitool” file in xbin. If the file is changed the permissions will need to be changed back. I Left in the old permissions in but commented them out in case a decision is made to change the file. Otherwise BT & wifi will not run at the same time.